requirements.txt
google-cloud-storage>=1.41.1
feedparser>=6.0.2
deepgram-sdk~=0.1.8
